Changed combat system
Changed the engage release system to a charge - release system. Now you
start charging whenever you want, and you have to release your attack in
the ReleaseBox of the enemies to hit them.
Added a list of hit accuracy to GUIScript that collects all the hits.
This should be used for score and kill streaks.